requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.7
sqlalchemy==2.0.32
openai==1.65.5
pytest==9.0.2
google-genai==0.5.0
python-dotenv==1.1.1